// src/lib/api/models/Image.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Image = {
    /**
     * Unique identifier for the image
     */
    imageId: string;
    /**
     * URL to the image
     */
    url: string;
    /**
     * Alternative text for the image
     */
    alt?: string | null;
};