// src/lib/api/models/StoryHead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StoryHead = {
    /**
     * Unique identifier for the story
     */
    storyId: string;
    /**
     * Name of the story
     */
    storyName: string;
    /**
     * Timestamp when the story was last edited
     */
    lastEdited: string;
    /**
     * URL to the cover image for the story
     */
    coverImage?: string | null;
};