// src/lib/api/models/UpdateImagesFromTextRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateImagesFromTextRequest = {
    /**
     * The Id of the user who owns the story
     */
    userId: string;
    /**
     * The Id of the story to update
     */
    storyId: string;
    /**
     * The updated text content to generate images from
     */
    updatedText: string;
};