// src/lib/api/models/UserDetailsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserDetailsResponse = {
    /**
     * Unique identifier for the user
     */
    userId: string;
    /**
     * Name of the user
     */
    userName: string;
    /**
     * Full name of the user
     */
    name?: string | null;
    /**
     * Date when the user created the account
     */
    accountCreated?: string | null;
};